+++ templates_c/fc27875d29dbedd3895ed63f733e0a057f87491e.file.orders.tpl.php
@@ -1,4 +1,4 @@
-<?php /* Smarty version Smarty-3.1.21-dev, created on 2015-03-25 13:29:14
+<?php /* Smarty version Smarty-3.1.21-dev, created on 2015-03-27 08:57:26
          compiled from "app\view\cms\orders.tpl" */ ?>
 <?php /*%%SmartyHeaderCode:1626555127b1d954023-03338267%%*/if(!defined('SMARTY_DIR')) exit('no direct access allowed');
 $_valid = $_smarty_tpl->decodeProperties(array (
@@ -7,7 +7,7 @@ $_valid = $_smarty_tpl->decodeProperties(array (
     'fc27875d29dbedd3895ed63f733e0a057f87491e' => 
     array (
       0 => 'app\\view\\cms\\orders.tpl',
-      1 => 1427286553,
+      1 => 1427442964,
       2 => 'file',
     ),
   ),
@@ -27,16 +27,9 @@ $_valid = $_smarty_tpl->decodeProperties(array (
   'has_nocache_code' => false,
 ),false); /*/%%SmartyHeaderCode%%*/?>
 <?php if ($_valid && !is_callable('content_55127b1d9a26d6_53501664')) {function content_55127b1d9a26d6_53501664($_smarty_tpl) {?><div class="container">
-	<div class="row vertical-offset-20">
-		<div class="col-md-2 pull-right">
-			<div class="panel panel-default">
-				<div class="panel-body text-centered">
-					<a class="btn btn-blue" href="/cms/"><span class="glyphicon glyphicon-search" aria-hidden="true"></span></a>
-				</div>
-			</div>
-		</div>
+	<div class="row cms vertical-offset-20">
 	</div>
-	<div class="row">
+	<div class="row cms">
 		<div class="col-md-12">
 			<div class="panel panel-default">
 				<div class="panel-heading text-centered">
@@ -70,9 +63,8 @@ $_smarty_tpl->tpl_vars['account']->_loop = true;
 						<?php } ?>
 						<td class="col-md-2 text-centered">
 							<a href="/cms/read_bestelling?id=<?php echo $_smarty_tpl->tpl_vars['order']->value->getId();?>
-" class="btn btn-primary"><span class="
-glyphicon glyphicon-sunglasses"></span></a>
-							<a class="btn btn-danger" href="/cms/delete_bestelling?id=<?php echo $_smarty_tpl->tpl_vars['order']->value->getId();?>
+" class="btn btn-primary"><span class="glyphicon glyphicon-sunglasses"></span></a>
+							<a onclick="return confirm('Weet je zeker dat je dit wilt verwijderen?');" class="btn btn-danger" href="/cms/delete_bestelling?id=<?php echo $_smarty_tpl->tpl_vars['order']->value->getId();?>
 "><span class="glyphicon glyphicon-trash" aria-hidden="true"></span></a>
 						</td>
 					</tr>
